fixed python code generation
